[Silver I] Title: 구간 합 구하기 5, Time: 1356 ms, Memory: 114880 KB -BaekjoonHub

+++ 백준/Silver/11660. 구간 합 구하기 5/구간 합 구하기 5.java
@@ -12,12 +12,12 @@ public class Main {
 		int N = Integer.parseInt(st.nextToken());	// 크기
 		int M = Integer.parseInt(st.nextToken());	// 합을 구해야 하는 횟수
 
-		long[][] arr = new long[N+1][N+1];
+		int[][] arr = new int[N+1][N+1];
 		
 		for(int i = 1; i <= N; i++) {
 			st = new StringTokenizer(br.readLine());	// N개의 수
 			for(int j = 1; j<= N; j++) {
-				arr[i][j] = arr[i][j-1] + Long.parseLong(st.nextToken());				
+				arr[i][j] = arr[i][j-1] + Integer.parseInt(st.nextToken());				
 			}
 		}
